docs(theme): add visual palette catalog
