+++ docs/src/content/docs/commands/charts.mdx
@@ -32,7 +32,7 @@ Every chart subcommand accepts the common data, grouping, and metadata flags:
 | `--output` | `-o` | *(stdout)* | Output file path. `.json` → JSON, else → HTML |
 | `--parser` | `-P` | `auto` | Parser: `auto` (detect), `go`, `js:tinybench`, `js:vitest`, `rs:criterion`, `rs:divan`, `csv`, `json` |
 | `--name` | `-n` | `Comparisons` | Dataset name |
-| `--theme` | | `default` | Initial series palette: one of the 13 built-ins or 2+ comma-separated `#rgb`/`#rrggbb` colors; see the [root command](/commands/root#flags) |
+| `--theme` | | `default` | Initial series palette: one of the 13 built-ins or 2+ comma-separated `#rgb`/`#rrggbb` colors; see [Color Themes](/ui/themes) |
 | `--description` | `-d` | `""` | Dataset description |
 | `--tag` | `-t` | `""` | Tag identifier for release tracking |
 | `--id` | | `""` | Stable dataset id for `?id=` deep links in the HTML UI |
