+++ ui/src/lib/themeCatalog.ts
@@ -1,0 +1,161 @@
+export const THEMES = {
+  default: [
+    '#5470C6',
+    '#3BA272',
+    '#FC8452',
+    '#73C0DE',
+    '#EE6666',
+    '#FAC858',
+    '#9A60B4',
+    '#EA7CCC',
+    '#91CC75',
+    '#FF9F7F',
+  ],
+  vintage: [
+    '#d87c7c',
+    '#919e8b',
+    '#d7ab82',
+    '#6e7074',
+    '#61a0a8',
+    '#efa18d',
+    '#787464',
+    '#cc7e63',
+    '#724e58',
+    '#4b565b',
+  ],
+  dark: [
+    '#dd6b66',
+    '#759aa0',
+    '#e69d87',
+    '#8dc1a9',
+    '#ea7e53',
+    '#eedd78',
+    '#73a373',
+    '#73b9bc',
+    '#7289ab',
+    '#91ca8c',
+  ],
+  westeros: [
+    '#516b91',
+    '#59c4e6',
+    '#edafda',
+    '#93b7e3',
+    '#a5e7f0',
+    '#cbb0e3',
+    '#3f5575',
+    '#41a7cb',
+    '#d58fc4',
+    '#789bc7',
+  ],
+  essos: [
+    '#893448',
+    '#d95850',
+    '#eb8146',
+    '#ffb248',
+    '#f2d643',
+    '#ebdba4',
+    '#6f293c',
+    '#b84445',
+    '#d96a39',
+    '#d0b72f',
+  ],
+  wonderland: [
+    '#4ea397',
+    '#22c3aa',
+    '#7bd9a5',
+    '#d0648a',
+    '#f58db2',
+    '#f2b3c9',
+    '#367d75',
+    '#199b89',
+    '#5dbd88',
+    '#ad4f75',
+  ],
+  walden: [
+    '#3fb1e3',
+    '#6be6c1',
+    '#626c91',
+    '#a0a7e6',
+    '#c4ebad',
+    '#96dee8',
+    '#318db5',
+    '#51b99b',
+    '#4e5674',
+    '#8188bc',
+  ],
+  chalk: [
+    '#fc97af',
+    '#87f7cf',
+    '#f7f494',
+    '#72ccff',
+    '#f7c5a0',
+    '#d4a4eb',
+    '#d2f5a6',
+    '#76f2f2',
+    '#ff7f9f',
+    '#5de0bd',
+  ],
+  infographic: [
+    '#C1232B',
+    '#27727B',
+    '#FCCE10',
+    '#E87C25',
+    '#B5C334',
+    '#FE8463',
+    '#9BCA63',
+    '#FAD860',
+    '#F3A43B',
+    '#60C0DD',
+  ],
+  macarons: [
+    '#2ec7c9',
+    '#b6a2de',
+    '#5ab1ef',
+    '#ffb980',
+    '#d87a80',
+    '#8d98b3',
+    '#e5cf0d',
+    '#97b552',
+    '#95706d',
+    '#dc69aa',
+  ],
+  roma: [
+    '#E01F54',
+    '#001852',
+    '#f5e8c8',
+    '#b8d2c7',
+    '#c6b38e',
+    '#a4d8c2',
+    '#f3d999',
+    '#d3758f',
+    '#dcc392',
+    '#2e4783',
+  ],
+  shine: [
+    '#c12e34',
+    '#e6b600',
+    '#0098d9',
+    '#2b821d',
+    '#005eaa',
+    '#339ca8',
+    '#cda819',
+    '#32a487',
+    '#d94a50',
+    '#f2c313',
+  ],
+  'purple-passion': [
+    '#8a7ca8',
+    '#e098c7',
+    '#8fd3e8',
+    '#71669e',
+    '#cc70af',
+    '#7cb4cc',
+    '#6d6189',
+    '#ba79aa',
+    '#72b8cc',
+    '#574e7d',
+  ],
+} as const satisfies Record<string, readonly string[]>
+
+export type ThemeName = keyof typeof THEMES
+export const THEME_NAMES = Object.keys(THEMES) as ThemeName[]
